--- Backend/.nyc_output/processinfo/b32a0d7f-f3bf-490a-8646-4e79e17da7b7.json
@@ -1,1 +1,0 @@
-{"parent":null,"pid":32696,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Users\\Rittr\\Desktop\\DAIICT\\Sem 5\\Software Engineering\\Final\\Backend\\node_modules\\mocha\\bin\\mocha.js","test/userController_test.js"],"execArgv":[],"cwd":"C:\\Users\\Rittr\\Desktop\\DAIICT\\Sem 5\\Software Engineering\\Final\\Backend","time":1732978046032,"ppid":53904,"coverageFilename":"C:\\Users\\Rittr\\Desktop\\DAIICT\\Sem 5\\Software Engineering\\Final\\Backend\\.nyc_output\\b32a0d7f-f3bf-490a-8646-4e79e17da7b7.json","externalId":"","uuid":"b32a0d7f-f3bf-490a-8646-4e79e17da7b7","files":["C:\\Users\\Rittr\\Desktop\\DAIICT\\Sem 5\\Software Engineering\\Final\\Backend\\controllers\\userController.js"]}
