+++ Backend/.nyc_output/processinfo/d11049b4-ef38-45f0-8f89-1d338bc7364b.json
@@ -1,0 +1,1 @@
+{"parent":"d4477f80-4a80-4675-96b5-010d9f793a02","pid":8064,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Users\\Rakshit\\Desktop\\Project\\IT314_G28_Restaurant_Reservation_System\\Backend\\node_modules\\mocha\\bin\\mocha.js"],"execArgv":[],"cwd":"C:\\Users\\Rakshit\\Desktop\\Project\\IT314_G28_Restaurant_Reservation_System\\Backend","time":1733052300820,"ppid":12952,"coverageFilename":"C:\\Users\\Rakshit\\Desktop\\Project\\IT314_G28_Restaurant_Reservation_System\\Backend\\.nyc_output\\d11049b4-ef38-45f0-8f89-1d338bc7364b.json","externalId":"","uuid":"d11049b4-ef38-45f0-8f89-1d338bc7364b","files":["C:\\Users\\Rakshit\\Desktop\\Project\\IT314_G28_Restaurant_Reservation_System\\Backend\\controllers\\restaurantController.js","C:\\Users\\Rakshit\\Desktop\\Project\\IT314_G28_Restaurant_Reservation_System\\Backend\\controllers\\userController.js"]}
