+++ Backend/.nyc_output/processinfo/d4477f80-4a80-4675-96b5-010d9f793a02.json
@@ -1,0 +1,1 @@
+{"parent":null,"pid":12448,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-cli.js","run","test"],"execArgv":[],"cwd":"C:\\Users\\Rakshit\\Desktop\\Project\\IT314_G28_Restaurant_Reservation_System\\Backend","time":1733052299811,"ppid":14600,"coverageFilename":"C:\\Users\\Rakshit\\Desktop\\Project\\IT314_G28_Restaurant_Reservation_System\\Backend\\.nyc_output\\d4477f80-4a80-4675-96b5-010d9f793a02.json","externalId":"","uuid":"d4477f80-4a80-4675-96b5-010d9f793a02","files":[]}
