+++ Backend/.nyc_output/processinfo/f1e1cd62-870f-4ba1-a79e-fd2a36646e4a.json
@@ -1,0 +1,1 @@
+{"parent":null,"pid":14276,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-cli.js","prefix","-g"],"execArgv":[],"cwd":"C:\\Users\\Rakshit\\Desktop\\Project\\IT314_G28_Restaurant_Reservation_System\\Backend","time":1733052298904,"ppid":13408,"coverageFilename":"C:\\Users\\Rakshit\\Desktop\\Project\\IT314_G28_Restaurant_Reservation_System\\Backend\\.nyc_output\\f1e1cd62-870f-4ba1-a79e-fd2a36646e4a.json","externalId":"","uuid":"f1e1cd62-870f-4ba1-a79e-fd2a36646e4a","files":[]}
