+++ Backend/stryker.restaurant.config.json
@@ -1,0 +1,23 @@
+{
+  "_comment": "This config was generated using 'stryker init'. Please take a look at: https://stryker-mutator.io/docs/stryker-js/configuration/ for more information.",
+  "packageManager": "npm",
+  "reporters": [
+    "html",
+    "clear-text",
+    "progress"
+  ],
+  "testRunner": "mocha",
+  "testRunner_comment": "Take a look at https://stryker-mutator.io/docs/stryker-js/mocha-runner for information about the mocha plugin.",
+  "coverageAnalysis": "perTest",
+  "mutate": [
+    "controllers/restaurantController.js"
+  ],
+  "mochaOptions": {
+    "spec": [
+      "test/restaurantController_test.js"
+    ]
+  },
+  "htmlReporter": {
+    "fileName": "reports/restaurant/index.html"
+  }
+}
